Further split basic and special purpose constraint tests
